Change all np.float to float
np.float has been a deprecated alias for float and is removed as of v1.24:
https://numpy.org/doc/stable/release/1.24.0-notes.html#expired-deprecations

+++ src/dials/algorithms/merging/merge.py
@@ -387,7 +387,7 @@ def merge_scaled_array_to_mtz_with_report_collection(
     if wavelength is None:
         wavelength = np.mean(
             np.array(
-                [expt.beam.get_wavelength() for expt in experiments], dtype=np.float
+                [expt.beam.get_wavelength() for expt in experiments], dtype=float
             )
         )
     with collect_html_data_from_merge() as collector:
